new language: go
This commit adds a new function to calculate prime numbers up to a given integer. The 'isPrime' function uses the Sieve of Eratosthenes algorithm to efficiently identify prime numbers. The main function has been updated to calculate the number of prime numbers between 0 and 8,999,999 and outputs the result along with the execution time.